pyproject.toml
[project]
name = "cortex-code-cost-analyzer"
version = "1.0.0"
description = "Streamlit app to analyze Cortex Code CLI credit consumption"
requires-python = ">=3.11"
dependencies = [
    "snowflake-connector-python>=3.3.0",
    "streamlit[snowflake]>=1.54.0",
]
